remove useless url setting and base_href helper
Sinatra offers an `uri` helper which is there to properly compute
URLs for the application. We now use it instead of computing the `base_href` by
hand. We also remove the `url` setting as system administrators should properly
configure their `config.ru` and proxy to allow Sinatra to compute the right URL
for the application.