implement file format 2.0
We want to be able to write files for which we will not be sure of the size
before all its content has been written. Easiest way to achieve this is two
have two files instead of one: the first for the encrypted file content, and
the second with metadata (in clear and encrypted, as before).

StoredFile is still able to read files with format 1.0. New files will be
written using the 2.0 format.