implement better AGPL compliance
We now detect if Coquelicot is running from a serviceable Git clone.
When it is the case, we offer to retrieve source using Git as we did
previously.

If there is a Git repository which is not usable, a warning is sent to the
logs.

In case source can't be provided by Git, we offer a link to an on-the-fly
created Gem that can be downloaded and unpacked. The version number of
the running software is mangled to add the server hostname and a date.