Switch to the default expiration time to one day
Usage as shown that Coquelicot was often used to exchange huge files that
might take quite some time to be uploaded. Sometimes even more than an hour.
In that case, the file would be immediately expired. Not the best user
experience.

So let's bump the default to one day. This should be less surprising to users.
Admins can still get back to the previous default of one hour using the
configuration file.