Add userpass authentication
The `userpass` authentication mechanism prompts for a user and password
to perform an upload. The credentials are stored as pairs of login/password in
the local configuration. Password are stored in an encrypted form using
bcrypt().

`userpass` configured with a single account can be used instead of `simplepass`
to allow users to make their browser retain the upload credentials.

Based on a patch from Rowan Thorpe.